requirements-dev.txt
-r requirements.txt

playwright==1.51.0
pytest==8.3.5
pytest-django==4.11.1
pytest-playwright==0.7.0
coverage==7.5.2
ruff==0.11.5